Developer yn

See merge request jeonghw4n/final-pjt!10

+++ final_project_back/articles/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_list_or_404, get_object_or_404
 
+from django.contrib.auth import get_user_model
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.decorators import api_view
@@ -27,17 +28,20 @@ def article_list(request):
 @api_view(['GET', 'DELETE', 'PUT'])
 def article_detail(request, pk):
     article = Article.objects.get(pk=pk)
+    user = request.user
     if request.method == "GET":
         serializer = ArticleSerializer(article)
         return Response(serializer.data)
     elif request.method == "DELETE":
-        article.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if user == article.user:
+            article.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
     elif request.method == 'PUT':
-        serializer = ArticleSerializer(article, data=request.data)
-        if serializer.is_valid(raise_exception=True):
-            serializer.save()
-            return Response(serializer.data)
+        if user == article.user:
+            serializer = ArticleSerializer(article, data=request.data)
+            if serializer.is_valid(raise_exception=True):
+                serializer.save()
+                return Response(serializer.data)
 
 
 @api_view(['GET'])
@@ -55,8 +59,10 @@ def comment_detail(request, pk):
         serializer = CommentSerializer(comment)
         return Response(serializer.data)
     elif request.method == "DELETE":
-        comment.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if request.user.is_authenticated:
+            if request.user == comment.user:
+                comment.delete()
+                return Response(status=status.HTTP_204_NO_CONTENT)
     elif request.method == 'PUT':
         serializer = CommentSerializer(comment, data=request.data)
         if serializer.is_valid(raise_exception=True):
@@ -74,11 +80,16 @@ def comment_create(request, article_pk):
         # serializer.save(article=article)
 
         return Response(serializer.data, status=status.HTTP_201_CREATED)
-    
-def likes(request, article_pk):
-    article = Article.objects.get(pk=article_pk)
-    user = request.user
-    if article.like_users.filter(id=user.id).exists():
-        article.like_users.remove(user)
-    else:
-        article.like_users.add(user)
+
+@api_view(['POST'])
+def likes(request, pk):
+    article = Article.objects.get(pk=pk)
+    user = get_user_model().objects.get(username=request.user)
+
+    if request.user.is_authenticated:
+        if article.like_users.filter(id=user.id).exists():
+            article.like_users.remove(user)
+        else:
+            article.like_users.add(user)
+        serializer = ArticleSerializer(article)
+        return Response(serializer.data)
